:heavy_plus_sign: :fire: Bruk motor fra kelvin felleskomponenter (#10)

+++ app/src/main/kotlin/no/nav/aap/behandlingsflyt/behandling/avklaringsbehov/BehandlingTilstandValidator.kt
@@ -14,7 +14,7 @@ class BehandlingTilstandValidator(connection: DBConnection) {
         val behandling = behandlingReferanseService.behandling(journalpostId)
         ValiderBehandlingTilstand.validerTilstandBehandling(behandling, behandlingVersjon)
 
-        val jobberForBehandling = flytJobbRepository.hentJobberForBehandling(behandling.id)
+        val jobberForBehandling = flytJobbRepository.hentJobberForBehandling(behandling.id.toLong())
         if (jobberForBehandling.isNotEmpty()) {
             throw BehandlingUnderProsesseringException()
         }
